fix(engine): easier that made things harder, and seven more the audit found (#207)
A full audit of the v3.0 engine — thirteen independent passes, then a round
of trying to knock each finding down. Eight survived; all eight are closed
here, each with a sweep that would have caught it.

- §41.1 A descent across a variation boundary landed in the journal
  unconditionally. The movement below is often trained one side at a time, so
  the same remembered dose is twice the work: 49 boundaries out of 49 got
  heavier, worst ×11.25. The landing now walks down from the journal until the
  plan fits the work already being done. Three boundaries still rise, at most
  ×2.00, and are named in the spec rather than left to be found.
- §41.2 Tapping "Done" on a probe set recorded nothing, so eight of the ten
  ladders were frozen for anyone who does not adjust numbers by hand.
- §41.3 The journal recorded the plan's top instead of the fact, because the
  rounded mean could not tell "took the top set" from "did not" — 2903
  inflated cells of 28880. The raw mean travels now; storage still snaps.
- §41.4 The probe gate reads the journal rather than the current dose.
- §41.5 Rhythm recognition widened from three gaps to eight: a regular but
  long cycle was read as a break every time, ending below where it started.
- §41.7 A pre-v3 state is migrated rather than discarded, reversing §40.8 —
  470 of 480 positions land no heavier, and a one-shot card on Today says so.
- §41.8 Five safety constants pinned by literal and by behaviour.
- Three of the six release gates did not run at all; the checklist is now
  machine-read by scripts/check_engine_gates.py, and its contract is in CI.

noHarder was tautological: it compared the landing against itself, so breaking
the rule produced zero failures. Block 13 gains a negative control that pins
the predicate, and the acceptance set is rebuilt for v3 — twenty blocks.

+++ DredfitTests/AppStoreTests+DebutBadge.swift
@@ -40,7 +40,7 @@ extension AppStoreTests {
             var day = 0
             // Every probe is passed: since §40.4 that is the only door into a
             // new movement, so a walk that ignored them would never get there.
-            func train(_ result: FeedbackResult, overrides: [Pattern: Int] = [:],
+            func train(_ result: FeedbackResult, overrides: [Pattern: Double] = [:],
                        skipped: Set<Pattern> = []) {
                 day += 1
                 let session = store.nextSession
